When running as pid1, propagate SIGTERM to pacemaker_remote
When pacemaker_remoted runs in a bundle as the container's main pid,
it runs into a special "pcmk-init" mode and forks a pacemaker_remoted
child.

When stopping the container, the container engine sends a SIGTERM to
the pcmk-init pid. Make sure that this one propagates the SIGTERM
to the real pacemaker_remoted pid, to give it a chance to stop
properly before the container engine times out and sends a SIGKILL.